requirements.txt
accelerate==0.34.2
aiofiles==23.2.1
aiohttp==3.10.5
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.0
dill==0.3.8
einops==0.8.0
fastapi==0.112.4
filelock==3.13.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.3
gradio==4.40.0
gradio_client==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.4
importlib_resources==6.4.5
Jinja2==3.1.3
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mmcv==1.7.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nltk==3.9.1
numpy==1.26.3
omegaconf==2.3.0
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.2.0
protobuf==3.20.2
pyarrow==17.0.0
pycocoevalcap==1.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rouge_score==0.1.2
safetensors==0.4.5
scipy==1.13.1
sentencepiece==0.1.99
timm==0.6.12
tokenizers==0.20.3
torchsde==0.2.6
tqdm==4.66.5
transformers==4.46.2
triton==3.1.0
xformers==0.0.28.post3